Change integer types

+++ dvd_info.c
@@ -68,7 +68,7 @@ struct dvd_audio {
 	char stream_id[DVD_AUDIO_STREAM_ID + 1];
 	char lang_code[DVD_AUDIO_LANG_CODE + 1];
 	char codec[DVD_AUDIO_CODEC + 1];
-	int channels;
+	uint8_t channels;
 };
 
 struct dvd_subtitle {
